% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alignment.R
\name{Alignment}
\alias{Alignment}
\title{Alignment object/class whatever they're called in R}
\description{
Simple wrapper around an alignment file (e.g., BAM)
for generating read coverage tracks, etc.
NOTE: called functions from this module are not yet implemented.
NOTE: use camelCase for R functions, snake_case for C++?
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Alignment-new}{\code{Alignment$new()}}
\item \href{#method-Alignment-getRawTrack}{\code{Alignment$getRawTrack()}}
\item \href{#method-Alignment-inferPairedEnd}{\code{Alignment$inferPairedEnd()}}
\item \href{#method-Alignment-inferFragmentLength}{\code{Alignment$inferFragmentLength()}}
\item \href{#method-Alignment-inferScaleFactor}{\code{Alignment$inferScaleFactor()}}
\item \href{#method-Alignment-clone}{\code{Alignment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Alignment-new"></a>}}
\if{latex}{\out{\hypertarget{method-Alignment-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alignment$new(
  filePath,
  assembly,
  pairedEnd = NULL,
  fragmentLength = NULL,
  scaleFactor = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Alignment-getRawTrack"></a>}}
\if{latex}{\out{\hypertarget{method-Alignment-getRawTrack}{}}}
\subsection{Method \code{getRawTrack()}}{
Return a 'raw' (unscaled) alignment-count track for a chromosome
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alignment$getRawTrack(chrom, binSize = 25L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chrom}}{chromosome name, e.g., \code{chr1}}

\item{\code{binSize}}{size of genomic intervals over which to measure aligned reads}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
numeric vector of counts per bin
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Alignment-inferPairedEnd"></a>}}
\if{latex}{\out{\hypertarget{method-Alignment-inferPairedEnd}{}}}
\subsection{Method \code{inferPairedEnd()}}{
Infer whether the file is paired-end by scanning a subset of reads.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alignment$inferPairedEnd(maxReads = 10000L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Alignment-inferFragmentLength"></a>}}
\if{latex}{\out{\hypertarget{method-Alignment-inferFragmentLength}{}}}
\subsection{Method \code{inferFragmentLength()}}{
Infer fragment length (paired or single-end).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alignment$inferFragmentLength()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Alignment-inferScaleFactor"></a>}}
\if{latex}{\out{\hypertarget{method-Alignment-inferScaleFactor}{}}}
\subsection{Method \code{inferScaleFactor()}}{
Determine scaling factor for normalization between replicates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alignment$inferScaleFactor(method = c("coverage", "RPKM", "DESeq"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Alignment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Alignment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alignment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
